Upload inital files

+++ src/main/java/com/qa/ims/controller/CrudController.java
@@ -14,6 +14,6 @@ public interface CrudController<T> {
 
 	T update();
 
-	int delete();
+	long delete();
 
 }
